+++ crates/rome_cli/src/execute/process_file.rs
@@ -104,6 +104,17 @@ pub(crate) fn process_file(ctx: &TraversalOptions, path: &Path) -> FileResult {
                 })
                 .and(
                     file_features
+                        .support_kind_for(&FeatureName::Format)
+                        .and_then(|support_kind| {
+                            if support_kind.is_not_enabled() {
+                                Some(support_kind)
+                            } else {
+                                None
+                            }
+                        }),
+                )
+                .and(
+                    file_features
                         .support_kind_for(&FeatureName::OrganizeImports)
                         .and_then(|support_kind| {
                             if support_kind.is_not_enabled() {
@@ -319,13 +330,7 @@ pub(crate) fn process_file(ctx: &TraversalOptions, path: &Path) -> FileResult {
             let should_write = match ctx.execution.traversal_mode() {
                 // In check mode do not run the formatter and return the result immediately,
                 // but only if the argument `--apply` is not passed.
-                TraversalMode::Check { .. } => {
-                    if ctx.execution.as_fix_file_mode().is_some() {
-                        true
-                    } else {
-                        return Ok(result);
-                    }
-                }
+                TraversalMode::Check { .. } => ctx.execution.as_fix_file_mode().is_some(),
                 TraversalMode::CI { .. } => false,
                 TraversalMode::Format { write, .. } => *write,
                 TraversalMode::Migrate { write: dry_run, .. } => *dry_run,